ORACLE-FORMS/automate-oracle-forms-gpt/automte-si3-v1/requirements.txt.txt
jpype1>=1.4.1
pillow
pytest
allure-pytest
opencv-python-headless
pytesseract
numpy
faker
pytest-html